suppressed some stupid undef. warnings, completed testing message policy deadlines and retries count. messages dropped by router socket handled correctly, yay\!

+++ include/cocaine/dealer/core/balancer.hpp
@@ -75,7 +75,7 @@ private:
 	std::vector<cocaine_endpoint> endpoints_;
 	boost::shared_ptr<cocaine::dealer::context> context_;
 	boost::shared_ptr<message_cache> message_cache_;
-	int current_endpoint_index_;
+	size_t current_endpoint_index_;
 };
 
 } // namespace dealer
